Fix: Users associated with notes

+++ app/Http/Controllers/TodoController.php
@@ -15,7 +15,7 @@ class TodoController extends Controller
     public function index()
     {
         //
-        return Todo::all();
+        return Todo::where('user_id',auth()->user()->id)->get();
     }
 
     /**
@@ -42,7 +42,11 @@ class TodoController extends Controller
             'completed' => 'required|boolean',
         ]);
 
-        $todo = Todo::create($data);
+        $todo = Todo::create([
+            'user_id' => auth()->user()->id,
+            'title' => $request->title,
+            'completed' => $request->completed
+        ]);
 
         return response($todo,201);
     }
@@ -79,6 +83,10 @@ class TodoController extends Controller
     public function update(Request $request, Todo $todo)
     {
         //
+        if($todo->user_id !== auth()->user()->id){
+            return response()->json('unauthorized!', 401);
+        }
+
         $data = $request->validate([
             'title' => 'required|string',
             'completed' => 'required|boolean',
@@ -102,7 +110,7 @@ class TodoController extends Controller
             'completed' => 'required|boolean',
         ]);
 
-        Todo::query()->update($data);
+        Todo::where('user_id', auth()->user()->id)->update($data);
 
         return response('Updated!',200);
     }
@@ -116,6 +124,10 @@ class TodoController extends Controller
     public function destroy(Todo $todo)
     {
         //
+        if($todo->user_id !== auth()->user()->id){
+            return response()->json('unauthorized!', 401);
+        }
+
         $todo->delete();
 
         return response('Deleted todo Item', 200);
@@ -130,12 +142,27 @@ class TodoController extends Controller
     public function destroyCompleted(Request $request)
     {
         //
+
+        $todosToDelete = $request->todos;
+
+        $userTodoIds = auth()->user()->todos->map(function($todo){
+            return $todo->id;
+        });
+
+        $valid = collect($todosToDelete)->every(function($value,$key) use ($userTodoIds){
+            return $userTodoIds->contains($value);
+        });
+
+        if(!$valid){
+            return response()->json('unauthorized!', 401);
+        }
+
         $request->validate([
             'todos' => 'required|array',
         ]);
 
         Todo::destroy($request->todos);
 
-        return response('Deleted', 200);
+        return response()->json('Deleted', 200);
     }
 }
